+++ shared/src/commonMain/kotlin/data/model/Account.kt
@@ -3,6 +3,6 @@ package data.model
 data class Account(
     val bankName: String,
     val personName: String,
-    val balance: Float,
+    val balance: Double,
     val lastTxs: List<Tx>
 )
